Move unrelated methods from the factory to helper
