+++ pkg/kubectl/cmd/util/factory_client_access.go
@@ -40,8 +40,6 @@ import (
 	restclient "k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
 	"k8s.io/kubernetes/pkg/api/legacyscheme"
-	api "k8s.io/kubernetes/pkg/apis/core"
-	"k8s.io/kubernetes/pkg/apis/extensions"
 	"k8s.io/kubernetes/pkg/client/clientset_generated/internalclientset"
 	"k8s.io/kubernetes/pkg/kubectl"
 	"k8s.io/kubernetes/pkg/kubectl/genericclioptions"
@@ -116,58 +114,6 @@ func (f *ring0Factory) RESTClient() (*restclient.RESTClient, error) {
 	}
 	setKubernetesDefaults(clientConfig)
 	return restclient.RESTClientFor(clientConfig)
-}
-
-func (f *ring0Factory) MapBasedSelectorForObject(object runtime.Object) (string, error) {
-	// TODO: replace with a swagger schema based approach (identify pod selector via schema introspection)
-	switch t := object.(type) {
-	case *api.ReplicationController:
-		return kubectl.MakeLabels(t.Spec.Selector), nil
-	case *api.Pod:
-		if len(t.Labels) == 0 {
-			return "", fmt.Errorf("the pod has no labels and cannot be exposed")
-		}
-		return kubectl.MakeLabels(t.Labels), nil
-	case *api.Service:
-		if t.Spec.Selector == nil {
-			return "", fmt.Errorf("the service has no pod selector set")
-		}
-		return kubectl.MakeLabels(t.Spec.Selector), nil
-	case *extensions.Deployment:
-		// TODO(madhusudancs): Make this smarter by admitting MatchExpressions with Equals
-		// operator, DoubleEquals operator and In operator with only one element in the set.
-		if len(t.Spec.Selector.MatchExpressions) > 0 {
-			return "", fmt.Errorf("couldn't convert expressions - \"%+v\" to map-based selector format", t.Spec.Selector.MatchExpressions)
-		}
-		return kubectl.MakeLabels(t.Spec.Selector.MatchLabels), nil
-	case *extensions.ReplicaSet:
-		// TODO(madhusudancs): Make this smarter by admitting MatchExpressions with Equals
-		// operator, DoubleEquals operator and In operator with only one element in the set.
-		if len(t.Spec.Selector.MatchExpressions) > 0 {
-			return "", fmt.Errorf("couldn't convert expressions - \"%+v\" to map-based selector format", t.Spec.Selector.MatchExpressions)
-		}
-		return kubectl.MakeLabels(t.Spec.Selector.MatchLabels), nil
-	default:
-		return "", fmt.Errorf("cannot extract pod selector from %T", object)
-	}
-}
-
-func (f *ring0Factory) ProtocolsForObject(object runtime.Object) (map[string]string, error) {
-	// TODO: replace with a swagger schema based approach (identify pod selector via schema introspection)
-	switch t := object.(type) {
-	case *api.ReplicationController:
-		return getProtocols(t.Spec.Template.Spec), nil
-	case *api.Pod:
-		return getProtocols(t.Spec), nil
-	case *api.Service:
-		return getServiceProtocols(t.Spec), nil
-	case *extensions.Deployment:
-		return getProtocols(t.Spec.Template.Spec), nil
-	case *extensions.ReplicaSet:
-		return getProtocols(t.Spec.Template.Spec), nil
-	default:
-		return nil, fmt.Errorf("cannot extract protocols from %T", object)
-	}
 }
 
 func (f *ring0Factory) SuggestedPodTemplateResources() []schema.GroupResource {
