+++ pkg/kubectl/cmd/util/factory_test.go
@@ -17,53 +17,10 @@ limitations under the License.
 package util
 
 import (
-	"strings"
 	"testing"
 
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/util/sets"
 	api "k8s.io/kubernetes/pkg/apis/core"
-	"k8s.io/kubernetes/pkg/kubectl"
-	"k8s.io/kubernetes/pkg/kubectl/genericclioptions"
 )
-
-func TestProtocolsForObject(t *testing.T) {
-	f := NewFactory(genericclioptions.NewTestConfigFlags())
-
-	pod := &api.Pod{
-		ObjectMeta: metav1.ObjectMeta{Name: "baz", Namespace: "test", ResourceVersion: "12"},
-		Spec: api.PodSpec{
-			Containers: []api.Container{
-				{
-					Ports: []api.ContainerPort{
-						{
-							ContainerPort: 101,
-							Protocol:      api.ProtocolTCP,
-						},
-						{
-							ContainerPort: 102,
-							Protocol:      api.ProtocolUDP,
-						},
-					},
-				},
-			},
-		},
-	}
-
-	expected := sets.NewString("101/TCP", "102/UDP")
-	protocolsMap, err := f.ProtocolsForObject(pod)
-	if err != nil {
-		t.Fatalf("Unexpected error: %v", err)
-	}
-
-	protocolsString := kubectl.MakeProtocols(protocolsMap)
-	protocolsStrings := strings.Split(protocolsString, ",")
-	got := sets.NewString(protocolsStrings...)
-
-	if !expected.Equal(got) {
-		t.Fatalf("Protocols mismatch! Expected %v, got %v", expected, got)
-	}
-}
 
 func TestMakePortsString(t *testing.T) {
 	tests := []struct {
