+++ pkg/kubectl/polymorphichelpers/BUILD
@@ -10,9 +10,11 @@ go_library(
         "historyviewer.go",
         "interface.go",
         "logsforobject.go",
+        "mapbasedselectorforobject.go",
         "objectpauser.go",
         "objectresumer.go",
         "portsforobject.go",
+        "protocolsforobject.go",
         "rollbacker.go",
         "statusviewer.go",
         "updatepodspec.go",
